move verovioAdapter.js from from `routers` into `utils`
